# docs/api/openapi-drf.yaml
openapi: 3.0.3
info:
  title: diconai API
  version: 1.0.0
  description: 산재 예방 통합 관제 시스템 — drf-server REST API
paths:
...
  /api/facilities/devices/{id}/:
    get:
      operationId: facilities_devices_retrieve
      summary: 전력 장비 상세 조회
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - Admin — Power Device
      security:
      - jwtAuth: []
      responses:
        '401':
          description: 인증 필요 (토큰 누락/만료)
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PowerDeviceAdminList'
          description: ''
        '404':
          description: 장치 없음
    put:
      operationId: facilities_devices_update
      summary: 전력 장비 수정 (Partial)
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - Admin — Power Device
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PowerDeviceAdminWriteRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PowerDeviceAdminWriteRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PowerDeviceAdminWriteRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '401':
          description: 인증 필요 (토큰 누락/만료)
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PowerDeviceAdminList'
          description: ''
        '400':
          description: 검증 실패
        '404':
          description: 장치 없음
    delete:
      operationId: facilities_devices_destroy
      summary: 전력 장비 비활성화 (Soft Delete)
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - Admin — Power Device
      security:
      - jwtAuth: []
      responses:
        '401':
          description: 인증 필요 (토큰 누락/만료)
        '204':
          description: 삭제 완료
        '404':
          description: 장치 없음
...
  /api/power-devices/{id}/:
    get:
      operationId: power_devices_retrieve
      summary: 전력 장비 상세 조회
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - Admin — Power Device
      security:
      - jwtAuth: []
      responses:
        '401':
          description: 인증 필요 (토큰 누락/만료)
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PowerDeviceAdminList'
          description: ''
        '404':
          description: 장치 없음
    put:
      operationId: power_devices_update
      summary: 전력 장비 수정 (Partial)
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - Admin — Power Device
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PowerDeviceAdminWriteRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PowerDeviceAdminWriteRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PowerDeviceAdminWriteRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '401':
          description: 인증 필요 (토큰 누락/만료)
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PowerDeviceAdminList'
          description: ''
        '400':
          description: 검증 실패
        '404':
          description: 장치 없음
    delete:
      operationId: power_devices_destroy
      summary: 전력 장비 비활성화 (Soft Delete)
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - Admin — Power Device
      security:
      - jwtAuth: []
      responses:
        '401':
          description: 인증 필요 (토큰 누락/만료)
        '204':
          description: 삭제 완료
        '404':
          description: 장치 없음
...
components:
  schemas:
...
    PowerDeviceAdminList:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        facility:
          type: integer
        facility_code:
          type: string
          readOnly: true
        facility_name:
          type: string
          readOnly: true
        facility_address:
          type: string
          readOnly: true
        device_id:
          type: string
          title: 하드웨어 식별자
          maxLength: 50
        device_name:
          type: string
          title: 사용자 정의 이름
          maxLength: 100
        manager_name:
          type: string
          readOnly: true
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - device_id
      - device_name
      - facility
      - facility_address
      - facility_code
      - facility_name
      - id
      - manager_name
    PowerDeviceAdminWriteRequest:
      type: object
      properties:
        facility:
          type: integer
        device_id:
          type: string
          minLength: 1
          title: 하드웨어 식별자
          maxLength: 50
        device_name:
          type: string
          minLength: 1
          title: 사용자 정의 이름
          maxLength: 100
        channel_count:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
          title: 채널 수
        is_active:
          type: boolean
      required:
      - device_id
      - device_name
      - facility
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT